updated steamdeck updater

+++ bash/steam/install_on_steam_deck.sh
@@ -8,11 +8,13 @@ source bash/util/package_util.sh
 
 if is_steam_deck; then
     # the point is to reinstall make glibc linux-api-headers sdl2, only if steam readonly was enabled (if steam deck updates)
-    echo "  > installing steam deck requirements"
     result=$(sudo steamos-readonly disable)
-    if [[ ! $result =~ "Nothing is performed" ]]; then
-        sudo pacman -S --noconfirm make glibc linux-api-headers sdl2 && make install-required
-        make && make install
+    if [[ ! $result =~ "already read-write" ]]; then
+        echo "  > installing steam deck requirements"
+        sudo pacman -S --noconfirm make glibc linux-api-headers sdl2
+        make install-required
+        make
+        make install
     else
         echo "  > steam deck was not updated"
     fi
